+++ example/main.js
@@ -1,27 +1,25 @@
-const React = require('react');
-const ReactDOM = require('react-dom');
-const MathJax = require('../src');
+/* @flow */
+import * as React from 'react';
+import ReactDOM from 'react-dom';
+import MathJax from '../src';
 
 const tex = `f(x) = \\int_{-\\infty}^\\infty
     \\hat f(\\xi)\\,e^{2 \\pi i \\xi x}
     \\,d\\xi`;
 
-const Example = React.createClass({
+class Example extends React.Component<*, *> {
     render() {
         return (
-            <MathJax.Context>
+            <MathJax.Provider>
                 <div>
-                    This is an inline math formula: <MathJax.Node inline>{'a = b'}</MathJax.Node>
+                    This is an inline math formula:{' '}
+                    <MathJax.Node inline formula={'a = b'} />
                     And a block one:
-
-                    <MathJax.Node>{tex}</MathJax.Node>
+                    <MathJax.Node formula={tex} />
                 </div>
-            </MathJax.Context>
+            </MathJax.Provider>
         );
     }
-});
+}
 
-ReactDOM.render(
-    <Example />,
-    document.getElementById('example')
-);
+ReactDOM.render(<Example />, document.getElementById('example'));