+++ HISTORY.md
@@ -2,6 +2,13 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+### [1.0.0]
+
+- Update for `react@16.4.0`
+- Rename `MathJax.Context` to `MathJax.Provider`
+- Load the MathJax script only if needed
+- Use `formula` prop instead of `childrenz` for `MathJax.Node`
+
 ### [0.1.1]
 
 - Accept `false` as value for `MathJaxContext#script`